Added commands, Changed project files
I deleted the .class files from the repo because we don't need them and
they are annoying to have there.
Also I added the commands from Dr. Wellington's notes and roughed in a
createUser command just to get a feel for how it will work. This has
brought a couple issues to my attention that I will address next time we
meet.